Allow to deploy freeIPA on nat-vm
This need is pushed by the tls-everywhere deploy, which should become
the default shortly.
It uses an IPA instance in order to get proper CA and certificates for
the different internal connections.

This change also affect the overall domain, presenting a new "domain"
parameter in order to get a proper FQDN on all the nodes.

Please note: enabling freeIPA service will modify the resources for the
nat-vm, pusing provisionned ram to 8G and 4 CPU cores.

Further work is needed on the undercloud and overcloud side in order to
fully enable the TLS-e "new generation" deploy.